Silence compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@23158 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/Properties/EditPropExternals.cpp
@@ -370,7 +370,7 @@ void CEditPropExternals::OnContextMenu(CWnd* /*pWnd*/, CPoint point)
     while (pos)
     {
         int index = m_ExtList.GetNextSelectedItem(pos);
-        if ((index >= 0)&&(index < m_externals.size()))
+        if ((index >= 0)&&(index < (int)m_externals.size()))
         {
             if (m_externals[index].headrev == SVN_INVALID_REVNUM)
             {
@@ -407,7 +407,7 @@ void CEditPropExternals::OnContextMenu(CWnd* /*pWnd*/, CPoint point)
                 {
                     int index = m_ExtList.GetNextSelectedItem(pos);
                     progDlg.SetProgress(count++, total);
-                    if ((index >= 0)&&(index < m_externals.size()))
+                    if ((index >= 0)&&(index < (int)m_externals.size()))
                     {
                         progDlg.SetLine(2, m_externals[index].url, true);
                         if (m_externals[index].headrev == SVN_INVALID_REVNUM)
@@ -439,7 +439,7 @@ void CEditPropExternals::OnContextMenu(CWnd* /*pWnd*/, CPoint point)
             while (pos)
             {
                 int index = m_ExtList.GetNextSelectedItem(pos);
-                if ((index >= 0)&&(index < m_externals.size()))
+                if ((index >= 0)&&(index < (int)m_externals.size()))
                 {
                     if (m_externals[index].headrev != SVN_INVALID_REVNUM)
                     {
